backend-ocr/requirements.txt
fastapi
uvicorn[standard]
pytesseract
pdf2image
Pillow
python-multipart